refactor: extract login overlay and form reusable mixins
