+++ packages/login/src/vaadin-login-form-mixin.js
@@ -1,0 +1,97 @@
+/**
+ * @license
+ * Copyright (c) 2018 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { LoginMixin } from './vaadin-login-mixin.js';
+
+/**
+ * @polymerMixin
+ * @mixes LoginMixin
+ */
+export const LoginFormMixin = (superClass) =>
+  class LoginFormMixin extends LoginMixin(superClass) {
+    static get observers() {
+      return ['_errorChanged(error)'];
+    }
+
+    /** @protected */
+    connectedCallback() {
+      super.connectedCallback();
+
+      if (!this.noAutofocus) {
+        this.$.vaadinLoginUsername.focus();
+      }
+    }
+
+    /** @private */
+    _errorChanged() {
+      if (this.error && !this._preventAutoEnable) {
+        this.disabled = false;
+      }
+    }
+
+    /**
+     * Submits the form.
+     */
+    submit() {
+      const userName = this.$.vaadinLoginUsername;
+      const password = this.$.vaadinLoginPassword;
+
+      if (this.disabled || !(userName.validate() && password.validate())) {
+        return;
+      }
+
+      this.error = false;
+      this.disabled = true;
+
+      const loginEventDetails = {
+        bubbles: true,
+        cancelable: true,
+        detail: {
+          username: userName.value,
+          password: password.value,
+        },
+      };
+
+      const firedEvent = this.dispatchEvent(new CustomEvent('login', loginEventDetails));
+      if (this.action && firedEvent) {
+        const csrfMetaName = document.querySelector('meta[name=_csrf_parameter]');
+        const csrfMetaValue = document.querySelector('meta[name=_csrf]');
+        if (csrfMetaName && csrfMetaValue) {
+          this.$.csrf.name = csrfMetaName.content;
+          this.$.csrf.value = csrfMetaValue.content;
+        }
+        this.querySelector('form').submit();
+      }
+    }
+
+    /** @protected */
+    _handleInputKeydown(e) {
+      if (e.key === 'Enter') {
+        const { currentTarget: inputActive } = e;
+        const nextInput =
+          inputActive.id === 'vaadinLoginUsername' ? this.$.vaadinLoginPassword : this.$.vaadinLoginUsername;
+        if (inputActive.validate()) {
+          if (nextInput.checkValidity()) {
+            this.submit();
+          } else {
+            nextInput.focus();
+          }
+        }
+      }
+    }
+
+    /** @protected */
+    _handleInputKeyup(e) {
+      const input = e.currentTarget;
+      if (e.key === 'Tab' && input instanceof HTMLInputElement) {
+        input.select();
+      }
+    }
+
+    /** @protected */
+    _onForgotPasswordClick() {
+      this.dispatchEvent(new CustomEvent('forgot-password'));
+    }
+  };
